Client_Project
This is done. Everything stated that it was valid until I added the favicon and it said it had issues with the formatting of what I had to paste in but we had to simply copy and paste it in so I cant change that it doesn't like that. I tried my absolute best to make this look as close as I could to the design. The course was really challenging for me, but I learned a lot. Thank you!